Close #1188

+++ pkgs/jnigen/test/simple_package_test/bindings/simple_package.dart
@@ -127,7 +127,7 @@ class Example_Nested extends jni.JObject {
   static final _new0 = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_NewObject")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -199,7 +199,7 @@ class Example_Nested extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallVoidMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -511,7 +511,7 @@ class Example extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallStaticVoidMethod")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_CallStaticVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -589,10 +589,10 @@ class Example extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_CallStaticIntMethod")
       .asFunction<
           jni.JniResult Function(
@@ -622,14 +622,14 @@ class Example extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_CallStaticIntMethod")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int,
@@ -684,7 +684,7 @@ class Example extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallVoidMethod")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -729,7 +729,7 @@ class Example extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallVoidMethod")
+                  ffi.VarArgs<(ffi.Uint8,)>)>>("globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -1086,7 +1086,7 @@ class Example extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
+                        ffi.Uint16,
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
@@ -1161,7 +1161,7 @@ class Example extends jni.JObject {
   static final _new1 = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_NewObject")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -1183,7 +1183,7 @@ class Example extends jni.JObject {
   static final _new2 = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64, ffi.Int64)>)>>("globalEnv_NewObject")
+                  ffi.VarArgs<(ffi.Int32, ffi.Uint8)>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int, int)>();
@@ -1210,8 +1210,8 @@ class Example extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
+                        ffi.Int32,
+                        ffi.Uint8,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_NewObject")
       .asFunction<
@@ -1245,14 +1245,14 @@ class Example extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int,
@@ -1307,7 +1307,7 @@ class Example extends jni.JObject {
   static final _addInts = ProtectedJniExtensions.lookup<
               ffi.NativeFunction<
                   jni.JniResult Function(ffi.Pointer<ffi.Void>,
-                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int64, ffi.Int64)>)>>(
+                      jni.JMethodIDPtr, ffi.VarArgs<(ffi.Int32, ffi.Int32)>)>>(
           "globalEnv_CallStaticIntMethod")
       .asFunction<
           jni.JniResult Function(
@@ -1452,7 +1452,7 @@ class Example extends jni.JObject {
                   jni.JThrowablePtr Function(
                       ffi.Pointer<ffi.Void>,
                       jni.JMethodIDPtr,
-                      ffi.VarArgs<(ffi.Int64, ffi.Pointer<ffi.Void>)>)>>(
+                      ffi.VarArgs<(ffi.Int32, ffi.Pointer<ffi.Void>)>)>>(
           "globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
@@ -1478,7 +1478,7 @@ class Example extends jni.JObject {
               jni.JThrowablePtr Function(
                   ffi.Pointer<ffi.Void>,
                   jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Int64,)>)>>("globalEnv_CallVoidMethod")
+                  ffi.VarArgs<(ffi.Int32,)>)>>("globalEnv_CallVoidMethod")
       .asFunction<
           jni.JThrowablePtr Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int)>();
@@ -3493,9 +3493,9 @@ class MyInterface<$T extends jni.JObject> extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
+                        ffi.Int32,
+                        ffi.Uint8,
+                        ffi.Uint16,
                         ffi.Double
                       )>)>>("globalEnv_CallLongMethod")
       .asFunction<
@@ -3762,9 +3762,9 @@ class MyInterfaceConsumer extends jni.JObject {
                       (
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
+                        ffi.Int32,
+                        ffi.Uint8,
+                        ffi.Uint16,
                         ffi.Double,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_CallStaticVoidMethod")
@@ -3822,9 +3822,9 @@ class MyInterfaceConsumer extends jni.JObject {
                       (
                         ffi.Pointer<ffi.Void>,
                         ffi.Pointer<ffi.Void>,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
+                        ffi.Int32,
+                        ffi.Uint8,
+                        ffi.Uint16,
                         ffi.Double,
                         ffi.Pointer<ffi.Void>
                       )>)>>("globalEnv_CallStaticVoidMethod")
@@ -4626,7 +4626,7 @@ class Exceptions extends jni.JObject {
   static final _new1 = ProtectedJniExtensions.lookup<
           ffi.NativeFunction<
               jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr,
-                  ffi.VarArgs<(ffi.Double,)>)>>("globalEnv_NewObject")
+                  ffi.VarArgs<(ffi.Float,)>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(
               ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, double)>();
@@ -4652,12 +4652,12 @@ class Exceptions extends jni.JObject {
                   jni.JMethodIDPtr,
                   ffi.VarArgs<
                       (
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64,
-                        ffi.Int64
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32,
+                        ffi.Int32
                       )>)>>("globalEnv_NewObject")
       .asFunction<
           jni.JniResult Function(ffi.Pointer<ffi.Void>, jni.JMethodIDPtr, int,
